Renamed events to future events
outsourced event preparation in utils.server
